# Conflicts:
#	composer.json
#	src/Plugin.php
#	src/controllers/PaymentsController.php
#	src/elements/Order.php
#	src/models/LineItem.php
#	src/services/Carts.php
#	src/web/assets/commercecp/dist/js/CommerceUpdateOrderStatusModal.js

+++ src/migrations/Install.php
@@ -209,7 +209,7 @@ class Install extends Migration
             'totalUses' => $this->integer()->notNull()->defaultValue(0)->unsigned(),
             'dateFrom' => $this->dateTime(),
             'dateTo' => $this->dateTime(),
-            'purchaseTotal' => $this->integer()->notNull()->defaultValue(0),
+            'purchaseTotal' => $this->decimal(14, 4)->notNull()->defaultValue(0),
             'purchaseQty' => $this->integer()->notNull()->defaultValue(0),
             'maxPurchaseQty' => $this->integer()->notNull()->defaultValue(0),
             'baseDiscount' => $this->decimal(14, 4)->notNull()->defaultValue(0),
